ImageFusion PWR 2024/requirements.txt
customtkinter==5.2.2
keras==3.3.3
matplotlib==3.8.4
numpy==1.24.3
opencv_python==4.10.0.84
Pillow==11.0.0
scikit_learn==1.5.2
skimage==0.0
tensorflow==2.16.1
tensorflow_intel==2.16.1
torch==2.3.1+cu121
